feat: Enhance notification and user services with new DTOs and refactor methods for improved type handling

+++ HomeHelperFinderAPI/Controllers/NotificationController.cs
@@ -16,10 +16,6 @@ public class NotificationController : ControllerBase
         _notificationService = notificationService;
     }
 
-    /// <summary>
-    /// Get all notifications with OData support
-    /// </summary>
-    /// <returns>List of notifications</returns>
     [HttpGet]
     [EnableQuery]
     public async Task<ActionResult<IEnumerable<NotificationDetailsDto>>> GetAll()
@@ -28,13 +24,8 @@ public class NotificationController : ControllerBase
         return Ok(notifications);
     }
 
-    /// <summary>
-    /// Get notification by ID
-    /// </summary>
-    /// <param name="id">Notification ID</param>
-    /// <returns>Notification details</returns>
     [HttpGet("{id:long}")]
-    public async Task<ActionResult<NotificationDetailsDto>> GetById(long id)
+    public async Task<ActionResult<NotificationDetailsDto>> GetById(int id)
     {
         if (!await _notificationService.ExistsAsync(id))
         {
@@ -45,11 +36,6 @@ public class NotificationController : ControllerBase
         return Ok(notification);
     }
 
-    /// <summary>
-    /// Create new notification
-    /// </summary>
-    /// <param name="createDto">Notification creation data</param>
-    /// <returns>Created notification</returns>
     [HttpPost]
     public async Task<ActionResult<NotificationDetailsDto>> Create([FromBody] NotificationCreateDto createDto)
     {
@@ -62,14 +48,8 @@ public class NotificationController : ControllerBase
         return CreatedAtAction(nameof(GetById), new { id = notification.NotificationId }, notification);
     }
 
-    /// <summary>
-    /// Update notification
-    /// </summary>
-    /// <param name="id">Notification ID</param>
-    /// <param name="updateDto">Notification update data</param>
-    /// <returns>Updated notification</returns>
     [HttpPut("{id:long}")]
-    public async Task<ActionResult<NotificationDetailsDto>> Update(long id, [FromBody] NotificationUpdateDto updateDto)
+    public async Task<ActionResult<NotificationDetailsDto>> Update(int id, [FromBody] NotificationUpdateDto updateDto)
     {
         if (!ModelState.IsValid)
         {
@@ -85,13 +65,8 @@ public class NotificationController : ControllerBase
         return Ok(notification);
     }
 
-    /// <summary>
-    /// Delete notification
-    /// </summary>
-    /// <param name="id">Notification ID</param>
-    /// <returns>Success status</returns>
     [HttpDelete("{id:long}")]
-    public async Task<ActionResult> Delete(long id)
+    public async Task<ActionResult> Delete(int id)
     {
         if (!await _notificationService.ExistsAsync(id))
         {
@@ -102,11 +77,6 @@ public class NotificationController : ControllerBase
         return Ok($"Notification with ID {id} has been deleted successfully.");
     }
 
-    /// <summary>
-    /// Get notifications by user ID
-    /// </summary>
-    /// <param name="userId">User ID</param>
-    /// <returns>List of user notifications</returns>
     [HttpGet("user/{userId:int}")]
     public async Task<ActionResult<IEnumerable<NotificationDetailsDto>>> GetByUserId(int userId)
     {
@@ -114,11 +84,6 @@ public class NotificationController : ControllerBase
         return Ok(notifications);
     }
 
-    /// <summary>
-    /// Get notifications by helper ID
-    /// </summary>
-    /// <param name="helperId">Helper ID</param>
-    /// <returns>List of helper notifications</returns>
     [HttpGet("helper/{helperId:int}")]
     public async Task<ActionResult<IEnumerable<NotificationDetailsDto>>> GetByHelperId(int helperId)
     {
@@ -126,11 +91,6 @@ public class NotificationController : ControllerBase
         return Ok(notifications);
     }
 
-    /// <summary>
-    /// Get unread notifications by user ID
-    /// </summary>
-    /// <param name="userId">User ID</param>
-    /// <returns>List of unread user notifications</returns>
     [HttpGet("user/{userId:int}/unread")]
     public async Task<ActionResult<IEnumerable<NotificationDetailsDto>>> GetUnreadByUserId(int userId)
     {
@@ -138,11 +98,6 @@ public class NotificationController : ControllerBase
         return Ok(notifications);
     }
 
-    /// <summary>
-    /// Get unread notifications by helper ID
-    /// </summary>
-    /// <param name="helperId">Helper ID</param>
-    /// <returns>List of unread helper notifications</returns>
     [HttpGet("helper/{helperId:int}/unread")]
     public async Task<ActionResult<IEnumerable<NotificationDetailsDto>>> GetUnreadByHelperId(int helperId)
     {
@@ -150,11 +105,6 @@ public class NotificationController : ControllerBase
         return Ok(notifications);
     }
 
-    /// <summary>
-    /// Get unread notification count by user ID
-    /// </summary>
-    /// <param name="userId">User ID</param>
-    /// <returns>Unread notification count</returns>
     [HttpGet("user/{userId:int}/unread-count")]
     public async Task<ActionResult<int>> GetUnreadCountByUserId(int userId)
     {
@@ -162,11 +112,6 @@ public class NotificationController : ControllerBase
         return Ok(count);
     }
 
-    /// <summary>
-    /// Get unread notification count by helper ID
-    /// </summary>
-    /// <param name="helperId">Helper ID</param>
-    /// <returns>Unread notification count</returns>
     [HttpGet("helper/{helperId:int}/unread-count")]
     public async Task<ActionResult<int>> GetUnreadCountByHelperId(int helperId)
     {
@@ -174,13 +119,8 @@ public class NotificationController : ControllerBase
         return Ok(count);
     }
 
-    /// <summary>
-    /// Mark notification as read
-    /// </summary>
-    /// <param name="id">Notification ID</param>
-    /// <returns>Success status</returns>
     [HttpPatch("{id:long}/mark-read")]
-    public async Task<ActionResult> MarkAsRead(long id)
+    public async Task<ActionResult> MarkAsRead(int id)
     {
         if (!await _notificationService.ExistsAsync(id))
         {
@@ -196,11 +136,6 @@ public class NotificationController : ControllerBase
         return BadRequest("Failed to mark notification as read.");
     }
 
-    /// <summary>
-    /// Mark all notifications as read for user
-    /// </summary>
-    /// <param name="userId">User ID</param>
-    /// <returns>Success status</returns>
     [HttpPatch("user/{userId:int}/mark-all-read")]
     public async Task<ActionResult> MarkAllAsReadByUserId(int userId)
     {
@@ -208,11 +143,6 @@ public class NotificationController : ControllerBase
         return Ok($"All notifications for user {userId} have been marked as read.");
     }
 
-    /// <summary>
-    /// Mark all notifications as read for helper
-    /// </summary>
-    /// <param name="helperId">Helper ID</param>
-    /// <returns>Success status</returns>
     [HttpPatch("helper/{helperId:int}/mark-all-read")]
     public async Task<ActionResult> MarkAllAsReadByHelperId(int helperId)
     {
